Refactor section component
