+++ src/containers/SignupForm/SignupForm.js
@@ -8,7 +8,6 @@ import { graphql, compose } from 'react-apollo'
 import Form from '../../components/Form'
 import Input from '../../components/Input'
 import Button from '../../components/Button'
-import Container from '../../components/Container'
 import { withRouter } from 'react-router-dom'
 import { GET_CURRENT_USER } from '../../query'
 import Alert from '../../components/Alert'
